requirements.txt
python-dotenv>=1.2.2
langchain>=0.3.24
langgraph>=1.1.3
langchain-openai>=0.3.14
langchain-core>=1.2.22
socksio>=1.0.0